rename block_cache component to block_cache_old
