+++ .gitignore
@@ -26,3 +26,4 @@ component/azstorage/logfile.txt
 logfile.txt
 **/logfile.txt
 **/logfile.txt.*
+*~
