+++ component/block_cache_old/blockpool_test.go
@@ -34,7 +34,7 @@
    SOFTWARE
 */
 
-package block_cache
+package block_cache_old
 
 import (
 	"math/rand"
