+++ component/block_cache_old/stream.go
@@ -31,7 +31,7 @@
    SOFTWARE
 */
 
-package block_cache
+package block_cache_old
 
 import (
 	"errors"
